+++ src/App.js
@@ -1,5 +1,5 @@
 import Pages from './pages/pages';
-import { Allwrap, Bg } from './styles/global/Allwrap';
+import { Allwrap, Bg } from './styles/global/allWrap';
 
 function App() {
   return (
