--- src/components/footer/footer.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const footer = () => {
-  return (
-    <div>footer</div>
-  )
-}
