+++ src/components/mainpost/mainsnspoststyle.jsx
