--- src/components/modal/snsModal.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const SnsModal = () => {
-  return (
-    <div>modal</div>
-  )
-}
