--- src/components/search/search.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const search = () => {
-  return (
-    <div>search</div>
-  )
-}
